# Make submode_vv_faster