A large source file mostly composed of  sequential operations over integers was taking ~300s to typecheck:
```
004.506s -pp
  000.254s parsing
    000.254s other
  289.355s typing
    289.014s infer
    000.003s check_sig
    000.002s save_cmi
    000.135s save_cmt
    000.201s other
  001.445s transl
  017.338s generate
    010.840s flambda2
      004.204s lambda_to_flambda
      006.636s flambda_to_cmm
    003.874s compile_phrases
      000.744s selection
      000.035s checkmach
        000.034s check zero_alloc
        000.001s other
      000.052s comballoc
      000.635s cse
      002.199s regalloc
        002.198s default
          000.257s liveness
          000.070s deadcode
          000.678s spill
          000.192s split
          000.816s regalloc
          000.181s mach to linear
            000.142s cfgize
              000.005s register_preds
              000.063s optimizations
              000.074s other
            000.038s cfg_to_linear
            000.001s other
          000.004s other
        000.001s other
      000.202s emit_fundecl
        000.202s emit
        000.001s other
      000.006s other
    000.394s write_asm
    000.651s compact
    001.106s assemble
    000.473s other
  000.004s other
000.023s other
```
After inspection with perf it seemed that ~80% of the time was spend in the call to `List.memq` within `Mode.submode_vv`.

Reordering the layout of fields in `'a var` to make `mvid` be the first element confirmed these numbers as it was enough to make typechecking go down to ~200s.
Going further and replacing the `list` by a `map` reduced typechecking time even further to a mere `7.4s`.


This PR makes the assumption that equality of vars can be reduced to equality of their unique identifier. If this is not the case then I think we can still use a Map but it will get trickier as vars are parametrized.