@lpw25 2023-11-03
I would mostly expect this to make most type-checking slower. I think these lists mostly shouldn't grow that large, so I'd be interested to see what pattern of code is managing to produce huge ones.