@riaqn 2023-11-03
The assumption is correct: physical equality of `var` can be tested by comparing their `mvid`.

I remember discussing using `Set` instead of `List` with @lpw25 , and we were uncertain about the overhead of `Set`. How does this PR compare against the `main` branch for normal code? (i.e., codes that doesn't benefit extremely from `VarMap`).

Speaking of which, why do we use `Map` instead of `Set`? Or are they equivalent?