@poechsel 2023-11-03
I don't have performance numbers yet for a full tree build but will have next week. Here the `Map` is needed instead of the `Set` has you still need to iterate over the different elements and a set can't have a `'a t` has the key.