@lpw25 2023-11-03
You can still use a `Set` by just defining a comparison function of vars that just uses their id. Not sure it would meaningfully affect the performance, but is probably makes the code easier to understand.