@lpw25 2023-11-03
Oh, sorry I misunderstood your point. You can still do it though, it "just" needs you to use recursive modules. It's almost [the classic example of them from the manual](https://v2.ocaml.org/manual/recursivemodules.html).