@lpw25 2023-11-03
No, that's still not enough. You can do it with the `Set`s in base/core using things like `Base.Comparator.Make1`, but it might not be doable using the stdlib's set.