@riaqn 2024-05-02
I think we might need to do similar things for the current newer mode solver, but we need to do some benchmark first to justify the  overhead. Also it's not gonna happen anytime soon.